Fix overwriting formatter selection
Symfony provides the possibility to give a default value for a choice type form field when setting up a form. The formatter type is such choice field.
It is currently however overwritten with the default formatter of the formatter pool without warning.
A common use case is to persist the format together with other settings of e.g. Sonata blocks and to re-edit them later in the same format, but currently the form field will just jump back to 'text' format and render the block content unusable.
Instead, the given value should only be overwritten if no valid format was given by the user.